Beginning to add support for attaching the RNA to the DNA

+++ src/uk/ac/ic/doc/protein_factory/Game.java
@@ -14,10 +14,13 @@ import android.view.SurfaceView;
 
 public class Game {
 	
-    private List<RNANucleotide> floatingRNA = new LinkedList<RNANucleotide>();
+    private Collection<RNANucleotide> floatingRNA = new LinkedList<RNANucleotide>();
     private Stack<RNANucleotide> unusedRNA = new Stack<RNANucleotide>();
     private List<DNANucleotide> backboneDNA = new LinkedList<DNANucleotide>();
     private static final int rnaCount = 20;
+    private static final int TOUCH_ACCURACY = 50; // px
+    private static final int SNAP_ACCURACY = 50; // px
+    public static final int SNAP_OFFSET = 85; // px
     protected Random gen = new Random();
     private Context c;
     
@@ -25,16 +28,13 @@ public class Game {
     	this.c = c;
 
 	    for (int i = 0; i < rnaCount; i++)
-	        floatingRNA.add(new RNANucleotide(c, gen));
+	        floatingRNA.add(new RNANucleotide(c, gen, this));
 	    
 	    for (int i = 0; i < displayWidth() / 50; i++)
-	        backboneDNA.add(new DNANucleotide(c,gen,i));
+	        backboneDNA.add(new DNANucleotide(c,i, this));
     }
     
-    private int displayWidth() {
-    	return c.getResources().getDisplayMetrics().widthPixels;
-    }
-
+    /* Called regularly by main loop */
     public void drawToCanvas(Canvas canvas)
     {
     	canvas.drawColor(Color.rgb(244, 235, 141));
@@ -44,17 +44,21 @@ public class Game {
         
         for (RNANucleotide rna : floatingRNA)
             rna.draw(canvas);
+
     }
 
-    protected void physics()
+    /* Called regularly by main loop */
+    public void physics()
     {
         Collection<RNANucleotide> rnaToKill = new Stack<RNANucleotide>();
-
+        
         for (RNANucleotide rna : floatingRNA)
         {
-       		rna.wobble(gen);
-            if (rna.getX() < 0) // TODO: Add offset so the whole image must be off the screen before we kill it
-              	rnaToKill.add(rna);
+        	if(!rna.touched()) {
+        		rna.wobbleLeft();
+        		if (rna.getX() < 0) // TODO: Add offset so the whole image must be off the screen before we kill it
+        			rnaToKill.add(rna);
+        	}
         }
 
         // Clean up RNA that went off the screen
@@ -65,37 +69,47 @@ public class Game {
         }
 
         for (DNANucleotide dna : backboneDNA)
-            dna.moveLeft();
+            dna.wobbleLeft();
     }
     
+    /* Can arrive at any time */
     public void touch(MotionEvent e) {
         if (e.getAction() == MotionEvent.ACTION_DOWN) {
-        	RNANucleotide closest = getClosest(e);
+        	RNANucleotide closest = (RNANucleotide) closestNucleotide((int)e.getX(), (int)e.getY(), TOUCH_ACCURACY, floatingRNA);
         	if(closest != null)
         		closest.setTouched(true);
         }
         if (e.getAction() == MotionEvent.ACTION_MOVE) {
         	for(RNANucleotide rna : floatingRNA) {
-        		if(rna.touched())
+        		if(rna.touched()) {
         			rna.move((int)e.getX(),(int)e.getY());
+        			attemptSnapToBackBone(rna);
+        		}
         	}
         }
         else if (e.getAction() == MotionEvent.ACTION_UP) {
         	for(RNANucleotide rna : floatingRNA) {
-        		rna.setTouched(false);
+        		if(rna.touched()) {
+        			rna.setTouched(false);
+       				attemptAttachToBackBone(rna);
+        		}
         	}        	
         }
     }
+
+    private int displayWidth() {
+    	return c.getResources().getDisplayMetrics().widthPixels;
+    }
     
-    public RNANucleotide getClosest (MotionEvent e)
+	private Nucleotide closestNucleotide (int x, int y, int maxdist, Collection<? extends Nucleotide> collection)
     {
-        RNANucleotide closest_rna = null;
-        int closest_dist = (35*35)*2; // Don't return anything further than 35 pixels away 
+        Nucleotide closest_rna = null;
+        int closest_dist = (maxdist*maxdist); // Don't return anything further than maxdist pixels away 
         							  // Also, magic numbers yay 
         int this_dist;
-        for (RNANucleotide rna : floatingRNA)
+        for (Nucleotide rna : collection)
         {
-            this_dist = rna.sqDist((int)e.getX(), (int)e.getY());
+            this_dist = rna.sqDist(x, y);
             if (this_dist < closest_dist)
             {
                 closest_dist = this_dist;
@@ -104,5 +118,18 @@ public class Game {
         }
 
         return closest_rna;
+    }
+	
+	
+    private void attemptSnapToBackBone(RNANucleotide rna) {
+    	DNANucleotide nearest = (DNANucleotide) closestNucleotide(rna.getX(), rna.getY()-SNAP_OFFSET, SNAP_ACCURACY, backboneDNA);
+    	if(nearest != null && !nearest.snapped())
+    		rna.move(nearest.getX()-2, nearest.getY()+SNAP_OFFSET); // getX()-2 is a hack - I think the images aren't aligned properly?
+    }
+	
+    private void attemptAttachToBackBone(RNANucleotide rna) {
+    	DNANucleotide nearest = (DNANucleotide) closestNucleotide(rna.getX(), rna.getY()-SNAP_OFFSET, SNAP_ACCURACY, backboneDNA);
+    	if(nearest != null)
+    		rna.snap(nearest, SNAP_OFFSET);
     }
 }
